file renaming: reference example
