+++ apps/docs/src/examples/extras/instanced-sprite/TreeSpriteAtlas.svelte
