+++ apps/docs/src/examples/extras/instanced-sprite/WalkingBehaviour.svelte
